% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{print "fars_map_state"}
\usage{
fars_map_state(state.num, year)
}
\description{
Draws a map of the specified state showing the reported accidents coordinates for the given year.
Throws an error if the state number isn't valid or if there are no accidents to plot.
Also doesn't work if provided parameters are vectors
}
\details{
@param year An integer or string indicating the year.

@param state.num An integer that represents the state number.

@return Plots a map of the state, where the fatalities are the dots on the map.

@example fars_map_state(30, 2013)
}
