% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Print "fars_read_years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years of integer or string that when the files were generated}
}
\value{
this function returns a list of data frames with MONTH and year columns
}
\description{
The function is to take a list of years and returns a list of data frames
with MONTH and year columns based on data in "accident_X.csv.bz2" files.
The files need to be in the working directory (Check function fars_read above!).
}
