% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsbydan-package.R
\docType{package}
\name{farsbydan-package}
\alias{farsbydan}
\alias{farsbydan-package}
\title{farsbydan: This package helps Analyzing States data}
\description{
Reading, summarizing, and mapping States Data.
}
\author{
\strong{Maintainer}: Dan-Angelus Zembere \email{danangelus97@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
